# backend/docs/openapi.yaml
openapi: 3.0.3
info:
  title: DP DB GATEWAY API
  version: "2.0"
  description: |
    数据库管理网关的 HTTP 接口。**每一条**已注册的路由都在本文件里 —— 由
    `TestOpenAPICoversEveryRoute` 静态守住:新增路由但没写进来,测试直接失败。

    ## 三套鉴权,不要混用

    | 面 | 前缀 | 鉴权 | 谁在调 |
    |---|---|---|---|
    | 控制台 | `/api/v1` | `Authorization: Bearer <JWT>` | 登录的人 |
    | 开放接口 | `/api/v1/open` | `Authorization: Bearer <key>.<secret>` | 外部系统 |
    | 回调 | `/api/v1/approvals/lark/callback` | 共享密钥(Bearer 或 `?secret=`) | 审批服务 |

    开放接口**不走**控制台的 IP 白名单:那个设置回答的是"哪些办公网可以打开控制台",
    把 CI runner 的出口段并进去等于为了一台机器放宽控制台的边界。每把凭据自带白名单。

    ## 响应包

    一律 HTTP 200,业务结果在 `code` 里 —— 包括错误。

    ```json
    { "code": 0, "msg": "ok", "data": { } }
    ```

    | code | 含义 | 客户端该做什么 |
    |---|---|---|
    | `0` | 成功 | 读 `data` |
    | `40001` | 参数错误 | `msg` 说明了哪里不对,直接展示 |
    | `40100` | 未登录 / Token 失效 / API 凭据无效 | 重新登录;开放接口检查 key.secret |
    | `40300` | 无权(菜单、能力矩阵、scope、非本人单据) | 展示 `msg`,不要重试 |
    | `40301` | 来源 IP 不在白名单 | 联系管理员放行 |
    | `42200` | 命令被拦截,已转审批 | `data.approvalNo` 是审批单号 |
    | `42600` | 脚本保存路径未配置 | 引导去【系统设置 · 网关】 |
    | `42601` | 导出保存路径未配置 | 同上 |
    | `42800` | 需要二次验证(TOTP) | 弹验证码输入,带 `mfaCode` 重试 |
    | `42900` | 登录尝试过于频繁 | 退避后再试 |
    | `50000` | 服务端错误 | 可重试;`msg` 不含内部细节 |

    ## 控制台的三层闸门

    每条控制台写接口要过:**菜单权限** → **能力矩阵(角色 × 能力 × 分层)** → **高危命令字典**。
    下面每个操作的 `summary` 都标了它要求的守卫,格式 `[菜单 · admin]`:

    - `[terminal]` 需要终端菜单;`[perms · admin]` 需要权限菜单**且**是平台管理员。
    - 没有标注 = 任何已登录用户可调。

...

...
servers:
  - url: /api/v1
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: 控制台会话 Token,由 `/auth/login` 签发。
    openApiKey:
      type: http
      scheme: bearer
      description: |
        开放接口凭据,格式 `<key>.<secret>`,在【系统设置 · 开放接口】签发。
        也接受 `X-Vela-Key` + `X-Vela-Secret` 两个请求头(给保留了 Authorization 的客户端)。
...
security:
  - bearerAuth: []
paths:
...
  /auth/login:
    post:
      tags: [auth]
      summary: 登录(受 IP 白名单与登录频控约束)
      security: []
      requestBody:
        required: true
        content: { application/json: { schema: { $ref: '#/components/schemas/LoginReq' } } }
      responses:
        '200':
          description: OK / 40100 凭据错误 / 42800 需 MFA 验证码 / 42900 过于频繁
          content: { application/json: { schema: { $ref: '#/components/schemas/R' } } }
...
  /notifications:
    get:
      tags: [auth]
      summary: 站内通知收件箱
      parameters:
        - { name: limit, in: query, schema: { type: integer, default: 30 } }
      responses: { '200': { $ref: '#/components/responses/Ok' } }
...
  /risk/check:
    post:
      tags: [terminal]
      summary: '[terminal] 三层预检,不执行'
      requestBody:
        content: { application/json: { schema: { $ref: '#/components/schemas/RiskCheckReq' } } }
      responses:
        '200':
          description: OK
          content: { application/json: { schema: { $ref: '#/components/schemas/RiskCheckResp' } } }
  /terminal/exec:
    post:
      tags: [terminal]
      summary: '[terminal] 执行命令(WS 不可用时的 REST 回退)'
      requestBody:
        content: { application/json: { schema: { $ref: '#/components/schemas/ExecReq' } } }
      responses:
        '200':
          description: OK / 42200 已转审批(返回 approvalNo)/ 42800 需二次验证 / 40300 拒绝
          content: { application/json: { schema: { $ref: '#/components/schemas/ExecResp' } } }
...
  /async-jobs/{id}:
    get:
      tags: [terminal]
      summary: '[terminal] 后台任务详情 + 流式日志'
      parameters: [{ name: id, in: path, required: true, schema: { type: integer } }]
      responses: { '200': { $ref: '#/components/responses/Ok' } }

  # ---------------------------------------------------------------- scripts
...
  /scripts/uploads/{id}/download:
    get:
      tags: [scripts]
      summary: '[terminal] 下载自己的脚本文件'
      parameters: [{ name: id, in: path, required: true, schema: { type: integer } }]
      responses: { '200': { description: 文件流 } }
  /scripts/uploads/{id}/content:
    get:
      tags: [scripts]
      summary: '[terminal] 读取脚本正文(在线查看)'
      parameters: [{ name: id, in: path, required: true, schema: { type: integer } }]
      responses: { '200': { $ref: '#/components/responses/Ok' } }
  /scripts/uploads/{id}:
    delete:
      tags: [scripts]
      summary: '[terminal] 删除自己的脚本'
      parameters: [{ name: id, in: path, required: true, schema: { type: integer } }]
      responses: { '200': { $ref: '#/components/responses/Ok' } }

  # ---------------------------------------------------------------- snippets
...
  /snippets/{id}:
    put:
      tags: [snippets]
      summary: '[terminal] 修改快捷脚本(占用已被绑定的槽位会释放原脚本)'
      parameters: [{ name: id, in: path, required: true, schema: { type: integer } }]
      responses: { '200': { $ref: '#/components/responses/Ok' } }
    delete:
      tags: [snippets]
      summary: '[terminal] 删除快捷脚本'
      parameters: [{ name: id, in: path, required: true, schema: { type: integer } }]
      responses: { '200': { $ref: '#/components/responses/Ok' } }

  # ---------------------------------------------------------------- export
...
  /export/download:
    get:
      tags: [export]
      summary: '[terminal] 下载导出分卷'
      parameters: [{ name: file, in: query, required: true, schema: { type: string } }]
      responses: { '200': { description: 文件流 } }

  # ---------------------------------------------------------------- connections
...
  /connections/{id}:
    put:
      tags: [connections]
      summary: '[db · admin] 编辑实例(留空口令表示保持不变)'
      parameters: [{ name: id, in: path, required: true, schema: { type: integer } }]
      responses: { '200': { $ref: '#/components/responses/Ok' } }
    patch:
      tags: [connections]
      summary: '[db · admin] 局部修改:status / tags / policy'
      parameters: [{ name: id, in: path, required: true, schema: { type: integer } }]
      responses: { '200': { $ref: '#/components/responses/Ok' } }
  /connections/{id}/test:
    post:
      tags: [connections]
      summary: '[db · admin] 测试连通性'
      parameters: [{ name: id, in: path, required: true, schema: { type: integer } }]
      responses: { '200': { $ref: '#/components/responses/Ok' } }
  /connections/{id}/schema:
    get:
      tags: [connections]
      summary: '[terminal] 库表树'
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
        - { name: database, in: query, schema: { type: string } }
      responses: { '200': { $ref: '#/components/responses/Ok' } }
  /connections/{id}/objects:
    get:
      tags: [connections]
      summary: '[terminal] 视图 / 函数 / 存储过程 / 触发器等对象'
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
        - { name: scope, in: query, schema: { type: string } }
        - { name: database, in: query, schema: { type: string }, description: PostgreSQL 系必传,目录按库隔离 }
      responses: { '200': { $ref: '#/components/responses/Ok' } }
  /connections/{id}/object-source:
    get:
      tags: [connections]
      summary: '[terminal] 单个对象的源码 / 建表 DDL'
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
        - { name: scope, in: query, schema: { type: string } }
        - { name: type, in: query, required: true, schema: { type: string } }
        - { name: name, in: query, required: true, schema: { type: string } }
        - { name: database, in: query, schema: { type: string } }
      responses: { '200': { $ref: '#/components/responses/Ok' } }
...
  /env-tiers/{code}:
    put:
      tags: [tiers]
      summary: '[envtier · admin] 修改分层(名称、MFA、危险横幅、扫描基准等)'
      parameters: [{ name: code, in: path, required: true, schema: { type: string } }]
      responses: { '200': { $ref: '#/components/responses/Ok' } }
    delete:
      tags: [tiers]
      summary: '[envtier · admin] 删除分层(仍有环境挂靠时拒绝)'
      parameters: [{ name: code, in: path, required: true, schema: { type: string } }]
      responses: { '200': { $ref: '#/components/responses/Ok' } }
...
  /environments/{code}:
    put:
      tags: [tiers]
      summary: '[envtier · admin] 修改环境(可改绑分层)'
      parameters: [{ name: code, in: path, required: true, schema: { type: string } }]
      responses: { '200': { $ref: '#/components/responses/Ok' } }
    delete:
      tags: [tiers]
      summary: '[envtier · admin] 删除环境;moveTo 必填,实例改挂过去而不是悬空'
      parameters: [{ name: code, in: path, required: true, schema: { type: string } }]
      requestBody:
        content: { application/json: { schema: { type: object, required: [moveTo], properties: { moveTo: { type: string } } } } }
      responses: { '200': { $ref: '#/components/responses/Ok' } }

  # ---------------------------------------------------------------- roles & users
...
  /roles/{id}:
    get:
      tags: [roles]
      summary: '[perms] 角色详情(能力矩阵 + 菜单 + 成员 + 标签)'
      parameters: [{ name: id, in: path, required: true, schema: { type: integer } }]
      responses: { '200': { $ref: '#/components/responses/Ok' } }
    patch:
      tags: [roles]
      summary: '[perms · admin] 改名称 / 描述 / 默认连接角色 / 可审批'
      parameters: [{ name: id, in: path, required: true, schema: { type: integer } }]
      responses: { '200': { $ref: '#/components/responses/Ok' } }
  /roles/{id}/menus:
    put:
      tags: [roles]
      summary: '[perms · admin] 设置菜单权限'
      parameters: [{ name: id, in: path, required: true, schema: { type: integer } }]
      responses: { '200': { $ref: '#/components/responses/Ok' } }
  /roles/{id}/capabilities:
    put:
      tags: [roles]
      summary: '[perms · admin] 设置能力矩阵(能力 × 分层 → allow/approve/deny)'
      description: |
        能力维度:`select` `write` `ddl` `grant` `conn` `approve` `explain` `release`。
        `release` 管的是**能不能发起发布单**,与语句本身能不能跑是两个问题。
      parameters: [{ name: id, in: path, required: true, schema: { type: integer } }]
      responses: { '200': { $ref: '#/components/responses/Ok' } }
  /roles/{id}/tags:
    put:
      tags: [roles]
      summary: '[perms · admin] 设置角色的数据访问范围(空 = 不限)'
      parameters: [{ name: id, in: path, required: true, schema: { type: integer } }]
      responses: { '200': { $ref: '#/components/responses/Ok' } }
  /roles/{id}/members:
    post:
      tags: [roles]
      summary: '[perms · admin] 添加成员'
      parameters: [{ name: id, in: path, required: true, schema: { type: integer } }]
      responses: { '200': { $ref: '#/components/responses/Ok' } }
  /roles/{id}/members/{userId}:
    delete:
      tags: [roles]
      summary: '[perms · admin] 移除成员'
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
        - { name: userId, in: path, required: true, schema: { type: integer } }
      responses: { '200': { $ref: '#/components/responses/Ok' } }
...
  /users/{id}:
    patch:
      tags: [users]
      summary: '[perms · admin] 改状态 / 角色(停用会立即作废其所有会话)'
      parameters: [{ name: id, in: path, required: true, schema: { type: integer } }]
      responses: { '200': { $ref: '#/components/responses/Ok' } }
  /users/{id}/password:
    post:
      tags: [users]
      summary: '[perms · admin] 重置密码(会作废该用户所有会话)'
      parameters: [{ name: id, in: path, required: true, schema: { type: integer } }]
      responses: { '200': { $ref: '#/components/responses/Ok' } }
  /users/{id}/mfa/reset:
    post:
      tags: [users]
      summary: '[perms · admin] 解绑该用户的 TOTP'
      parameters: [{ name: id, in: path, required: true, schema: { type: integer } }]
      responses: { '200': { $ref: '#/components/responses/Ok' } }
  /users/{id}/mfa/bind:
    post:
      tags: [users]
      summary: '[perms · admin] 代为生成并绑定 TOTP(返回密钥与 otpauth URI)'
      parameters: [{ name: id, in: path, required: true, schema: { type: integer } }]
      responses: { '200': { $ref: '#/components/responses/Ok' } }
  /users/{id}/tags:
    get:
      tags: [users]
      summary: '[perms] 该用户的数据访问范围'
      parameters: [{ name: id, in: path, required: true, schema: { type: integer } }]
      responses: { '200': { $ref: '#/components/responses/Ok' } }
    put:
      tags: [users]
      summary: '[perms · admin] 设置个人范围 —— 覆盖角色范围,清空则回落到角色'
      parameters: [{ name: id, in: path, required: true, schema: { type: integer } }]
      responses: { '200': { $ref: '#/components/responses/Ok' } }

  # ---------------------------------------------------------------- risk rules
...
  /risk-commands/{name}:
    patch:
      tags: [risk-rules]
      summary: '[rules · admin] 改单个分层的等级'
      parameters: [{ name: name, in: path, required: true, schema: { type: string } }]
      responses: { '200': { $ref: '#/components/responses/Ok' } }
    delete:
      tags: [risk-rules]
      summary: '[rules · admin] 删除命令(删掉等于全分层放行)'
      parameters: [{ name: name, in: path, required: true, schema: { type: string } }]
      responses: { '200': { $ref: '#/components/responses/Ok' } }

  # ---------------------------------------------------------------- sql review
...
  /sql-review/rules/{id}:
    put:
      tags: [sql-review]
      summary: '[rules · admin] 改级别 / 开关 / 参数 / 文案;内置规则的编码与方言不可改'
      parameters: [{ name: id, in: path, required: true, schema: { type: integer } }]
      responses: { '200': { $ref: '#/components/responses/Ok' } }
    delete:
      tags: [sql-review]
      summary: '[rules · admin] 删除自定义规则 —— 内置规则只能停用'
      parameters: [{ name: id, in: path, required: true, schema: { type: integer } }]
      responses: { '200': { description: 'OK / 40001 内置规则不可删除' } }
...
  /sql-review/check:
    post:
      tags: [sql-review]
      summary: '[terminal] 规范审查(connectionId 优先于 dialect:目标实例决定方言)'
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required: [sql]
              properties:
                connectionId: { type: integer, format: int64 }
                dialect: { type: string, enum: [mysql, tidb, dws, oracle] }
                sql: { type: string }
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  instance: { type: string }
                  result: { $ref: '#/components/schemas/ReviewResult' }

  # ---------------------------------------------------------------- pipelines
  /pipelines:
    get:
      tags: [pipelines]
      summary: '[pipeline] 发布流程模板(含阶段)'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema: { type: array, items: { $ref: '#/components/schemas/PipelineDetail' } }
    post:
      tags: [pipelines]
      summary: '[pipeline · admin] 新建流程'
      requestBody:
        content: { application/json: { schema: { $ref: '#/components/schemas/PipelineDetail' } } }
      responses: { '200': { $ref: '#/components/responses/Ok' } }
  /pipelines/{id}:
    get:
      tags: [pipelines]
      summary: '[pipeline] 单个流程'
      parameters: [{ name: id, in: path, required: true, schema: { type: integer } }]
      responses: { '200': { $ref: '#/components/responses/Ok' } }
    put:
      tags: [pipelines]
      summary: '[pipeline · admin] 覆盖式保存 —— 阶段整体替换,运行中的发布单不受影响'
      parameters: [{ name: id, in: path, required: true, schema: { type: integer } }]
      responses: { '200': { $ref: '#/components/responses/Ok' } }
    delete:
      tags: [pipelines]
      summary: '[pipeline · admin] 删除流程(历史发布单保留自己的阶段快照)'
      parameters: [{ name: id, in: path, required: true, schema: { type: integer } }]
      responses: { '200': { $ref: '#/components/responses/Ok' } }
  /releases:
    get:
      tags: [pipelines]
      summary: '[pipeline] 发布单分页'
      parameters:
        - { name: scope, in: query, schema: { type: string, enum: [mine, all], default: mine }, description: all 需监管角色 }
        - { name: status, in: query, schema: { type: string } }
        - { name: page, in: query, schema: { type: integer, default: 1 } }
        - { name: pageSize, in: query, schema: { type: integer, default: 20, maximum: 100 } }
      responses:
        '200':
          description: OK。列表里阶段的 log/findings 为空,详情接口才带。
          content:
            application/json:
              schema:
                type: object
                properties:
                  items: { type: array, items: { $ref: '#/components/schemas/ReleaseView' } }
                  total: { type: integer }
                  page: { type: integer }
                  pageSize: { type: integer }
    post:
      tags: [pipelines]
      summary: '[pipeline] 提交发布'
      description: |
        提交时按当前角色与实例**复判一次**:命中拒绝直接 40300;需审批的变更必须选含
        审批阶段的流程,否则拒收(不建单)。执行阶段会再判一次 —— 等待期间规则可能变。
      requestBody:
        content: { application/json: { schema: { $ref: '#/components/schemas/ReleaseReq' } } }
      responses:
        '200':
          description: OK / 40300 能力矩阵或 release 能力拒绝 / 42800 需二次验证 / 40001 流程不含审批阶段
          content: { application/json: { schema: { $ref: '#/components/schemas/ReleaseView' } } }
  /releases/{id}:
    get:
      tags: [pipelines]
      summary: '[pipeline] 发布单详情(各阶段状态、日志、审查结果)'
      parameters: [{ name: id, in: path, required: true, schema: { type: integer } }]
      responses: { '200': { $ref: '#/components/responses/Ok' } }
  /releases/{id}/abort:
    post:
      tags: [pipelines]
      summary: '[pipeline] 终止发布 —— 仅 pending/waiting;running 不可终止'
      description: 执行中的发布可能已经打到目标库,一个"已终止"的假承诺比没有这个按钮更糟。
      parameters: [{ name: id, in: path, required: true, schema: { type: integer } }]
      responses: { '200': { $ref: '#/components/responses/Ok' } }
  /releases/{id}/stages/{stageId}/continue:
    post:
      tags: [pipelines]
      summary: '[pipeline] 人工确认阶段继续(发起人不能确认自己的发布)'
      parameters:
        - { name: id, in: path, required: true, schema: { type: integer } }
        - { name: stageId, in: path, required: true, schema: { type: integer } }
      responses: { '200': { $ref: '#/components/responses/Ok' } }

  # ---------------------------------------------------------------- approvals
  /approvals:
    get:
      tags: [approvals]
      summary: '[approve] 审批单分页'
      parameters:
        - { name: scope, in: query, schema: { type: string, enum: [mine, all] } }
        - { name: ap, in: query, schema: { type: string }, description: 按单号精确查一张,不受分页影响 }
        - { name: page, in: query, schema: { type: integer } }
        - { name: pageSize, in: query, schema: { type: integer } }
      responses: { '200': { $ref: '#/components/responses/Ok' } }
...